Update tests to detect bad isSafe() implementation
Some implementaitons of isSafe() have a bad bug.
They are passing the original 4 unit/system tests of isSafe(), but
only because they are counting the number of times find candidate
returns a successful result.  But if find candidate returns
no candidate after 1 or more processes complete, but not all of them,
then this new test should catch bad implementations that don't
detect correctly when no more candidates are available to
run.